[CHORE] Refactor fruit components
