+++ src/features/island/fruit/Soil.tsx
@@ -1,189 +1,48 @@
-import React from "react";
+import React, { useState } from "react";
 
-import classNames from "classnames";
 import soil from "assets/land/soil2.png";
-
-import { getTimeLeft, secondsToString } from "lib/utils/time";
-
-import { ProgressBar } from "components/ui/ProgressBar";
-import { InnerPanel } from "components/ui/Panel";
-import { addNoise } from "lib/images";
-import { PlantedFruit } from "features/game/types/game";
+import selectBox from "assets/ui/select/select_box.png";
+import classNames from "classnames";
 import { PIXEL_SCALE } from "features/game/lib/constants";
-import useUiRefresher from "lib/utils/hooks/useUiRefresher";
-import { FRUIT, FruitName } from "features/game/types/fruits";
-import { FruitLifecycle, FRUIT_LIFECYCLE } from "./fruits";
-import deadTree from "assets/fruit/dead_tree.webp";
 
 interface Props {
-  plantedFruit?: PlantedFruit;
-  showFruitDetails?: boolean;
-  showTimers: boolean;
+  playing: boolean;
+  onClick: () => void;
 }
 
-const CROP_NOISE_LEVEL = 0.1;
+export const Soil: React.FC<Props> = ({ playing, onClick }) => {
+  const [showSelectBox, setShowSelectBox] = useState(false);
 
-const getFruitImage = (imageSource: any): JSX.Element => {
-  return (
-    <img
-      className="relative"
-      style={{
-        bottom: "19px",
-        zIndex: "1",
-      }}
-      src={imageSource}
-      onLoad={(e) => addNoise(e.currentTarget, CROP_NOISE_LEVEL)}
-    />
-  );
-};
-
-const Timer = ({
-  percentage,
-  seconds,
-}: {
-  percentage: number;
-  seconds: number | undefined;
-}) => {
   return (
     <div
-      className="absolute"
-      style={{
-        top: `${PIXEL_SCALE * 24.2}px`,
-        width: `${PIXEL_SCALE * 15}px`,
-      }}
+      className="flex justify-center"
+      onMouseEnter={() => setShowSelectBox(true)}
+      onMouseLeave={() => setShowSelectBox(false)}
     >
-      <ProgressBar
-        percentage={percentage}
-        seconds={seconds}
-        type={"progress"}
-        formatLength="short"
-      />
-    </div>
-  );
-};
-
-const Popover = ({
-  showFruitDetails,
-  lifecycle,
-  plantedFruitName,
-  timeLeft,
-}: {
-  showFruitDetails: boolean | undefined;
-  lifecycle: FruitLifecycle;
-  plantedFruitName: FruitName;
-  timeLeft: number;
-}) => {
-  return (
-    <InnerPanel
-      className={classNames(
-        "ml-10 transition-opacity absolute whitespace-nowrap sm:opacity-0 -bottom-2 w-fit left-1 z-50 pointer-events-none",
-        {
-          "opacity-100": showFruitDetails,
-          "opacity-0": !showFruitDetails,
-        }
-      )}
-    >
-      <div className="flex flex-col text-xxs text-white text-shadow ml-2 mr-2">
-        <div className="flex flex-1 items-center justify-center">
-          <img src={lifecycle.ready} className="w-4 mr-1" />
-          <span>{plantedFruitName}</span>
-        </div>
-        <span className="flex-1">
-          {secondsToString(timeLeft, { length: "medium" })}
-        </span>
-      </div>
-    </InnerPanel>
-  );
-};
-
-export const Soil: React.FC<Props> = ({
-  plantedFruit,
-  showFruitDetails,
-  showTimers,
-}) => {
-  useUiRefresher({ active: !!plantedFruit });
-
-  if (!plantedFruit) {
-    return getFruitImage(soil);
-  }
-
-  const { harvestSeconds } = FRUIT()[plantedFruit.name];
-  const lifecycle = FRUIT_LIFECYCLE[plantedFruit.name];
-
-  if (!plantedFruit.harvestsLeft) {
-    return (
       <img
         className="relative"
         style={{
-          bottom: "10px",
+          bottom: "9px",
           zIndex: "1",
-          height: `${PIXEL_SCALE * 25}px`,
+          width: `${PIXEL_SCALE * 16}px`,
+          height: `${PIXEL_SCALE * 26}px`,
         }}
-        src={deadTree}
-        onLoad={(e) => addNoise(e.currentTarget, CROP_NOISE_LEVEL)}
+        src={soil}
       />
-    );
-  }
-
-  // weather the tree is replinishing
-  if (plantedFruit.harvestedAt) {
-    const replenishingTimeLeft = getTimeLeft(
-      plantedFruit.harvestedAt,
-      harvestSeconds
-    );
-    if (replenishingTimeLeft > 0) {
-      const replenishPercentage =
-        100 - (replenishingTimeLeft / harvestSeconds) * 100;
-      return (
-        <div className="relative w-full h-full flex justify-center">
-          {getFruitImage(lifecycle.harvested)}
-          {showTimers && (
-            <Timer
-              percentage={replenishPercentage}
-              seconds={replenishingTimeLeft}
-            />
-          )}
-          <Popover
-            showFruitDetails={showFruitDetails}
-            lifecycle={lifecycle}
-            plantedFruitName={plantedFruit.name}
-            timeLeft={replenishingTimeLeft}
-          />
-        </div>
-      );
-    }
-  }
-
-  // Seedling
-  const growingTimeLeft = getTimeLeft(plantedFruit.plantedAt, harvestSeconds);
-  if (growingTimeLeft > 0) {
-    const growPercentage = 100 - (growingTimeLeft / harvestSeconds) * 100;
-    const isAlmostReady = growPercentage >= 50;
-    const isHalfway = growPercentage >= 25 && !isAlmostReady;
-
-    return (
-      <div className="relative w-full h-full flex justify-center">
-        {getFruitImage(
-          isAlmostReady
-            ? lifecycle.almost
-            : isHalfway
-            ? lifecycle.halfway
-            : lifecycle.seedling
-        )}
-
-        {showTimers && (
-          <Timer percentage={growPercentage} seconds={growingTimeLeft} />
-        )}
-
-        <Popover
-          showFruitDetails={showFruitDetails}
-          lifecycle={lifecycle}
-          plantedFruitName={plantedFruit.name}
-          timeLeft={growingTimeLeft}
+      {playing && (
+        <img
+          src={selectBox}
+          className={classNames("absolute z-40 cursor-pointer", {
+            "opacity-100": showSelectBox,
+            "opacity-0": !showSelectBox,
+          })}
+          style={{
+            top: "21px",
+            width: `${PIXEL_SCALE * 16}px`,
+          }}
+          onClick={() => onClick()}
         />
-      </div>
-    );
-  }
-
-  return getFruitImage(lifecycle.ready);
+      )}
+    </div>
+  );
 };
